+++ projects/fitness-ledger/mini_program/miniprogram/config/env.example.js
@@ -1,0 +1,4 @@
+module.exports = {
+  envId: "",
+  staleAfterHours: 48
+};